+++ ISV/include/isv_worker.h
@@ -122,7 +122,7 @@ class ISVWorker : public RefBase
 
         // alloc/free VA surface
         status_t allocSurface(uint32_t* width, uint32_t* height,
-                uint32_t stride, uint32_t format, uint32_t handle, int32_t* surfaceId);
+                uint32_t stride, uint32_t format, unsigned long handle, int32_t* surfaceId);
         status_t freeSurface(int32_t* surfaceId);
 
         ISVWorker();
